feat: move ga code to PageHead
